+++ clinica-medica-comum/src/main/java/br/edu/imepac/comum/services/PacienteService.java
@@ -29,7 +29,7 @@ public class PacienteService {
         return modelMapper.map(paciente, PacienteDto.class);
     }
 
-    public PacienteDto atualizarPaciente(Long id, PacienteDto dto) {
+    public PacienteDto atualizarPaciente(Integer id, PacienteDto dto) {
         log.info("Atualizando paciente com ID: {}", id);
         Paciente existente = repository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Paciente não encontrado com ID: " + id));
@@ -38,14 +38,14 @@ public class PacienteService {
         return modelMapper.map(atualizado, PacienteDto.class);
     }
 
-    public void removerPaciente(Long id) {
+    public void removerPaciente(Integer id) {
         log.info("Removendo paciente com ID: {}", id);
         Paciente paciente = repository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Paciente não encontrado com ID: " + id));
         repository.delete(paciente);
     }
 
-    public PacienteDto buscarPorId(Long id) {
+    public PacienteDto buscarPorId(Integer id) {
         log.info("Buscando paciente com ID: {}", id);
         Paciente paciente = repository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Paciente não encontrado com ID: " + id));
